backend/requirements.txt
djangorestframework
django-cors-headers
django-nested-admin
django-modeltranslation
django-deep-translator
django-filter
django-redis
drf-yasg
python-decouple
psycopg2-binary
pillow
celery
redis
msgpack
gunicorn
pycryptodome
djangorestframework-simplejwt
drf-spectacular
django-axes
django-password-validators
django-prometheus
docxtpl
django-modeltranslation
zeep
lxml
django-tinymce
qrcode
pypandoc
pandas
openpyxl